project/backend/requirements.txt
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
Pillow==10.2.0
python-dotenv==1.0.1
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv